Исправление тестов

+++ src/main/java/ru/yandex/practicum/filmorate/mapper/FilmRowMapper.java
@@ -15,7 +15,7 @@ public class FilmRowMapper implements RowMapper<Film> {
     @Override
     public Film mapRow(ResultSet rs, int rowNum) throws SQLException {
         Film film = new Film();
-        film.setId(rs.getInt("id"));
+        film.setId(rs.getLong("id"));
         film.setName(rs.getString("name"));
         film.setDescription(rs.getString("description"));
         film.setReleaseDate(rs.getObject("release_date", LocalDate.class));
